requirements.txt
beautifulsoup4==4.11.2
blinker==1.5
certifi==2022.12.7
chardet==3.0.4
click==8.1.3
Flask==2.0.0
Flask-Admin==1.5.4
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-QRcode==3.1.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
idna==2.8
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.0.0
Markdown==3.1.1
MarkupSafe==2.1.1
numpy==1.21.6
pandas==1.3.5
Pillow==9.4.0
protobuf==3.20.1
pypng==0.20220715.0
PyQRCode==1.2.1
python-dateutil==2.8.2
pytz==2022.7.1
qrcode==7.3.1
requests==2.22.0
six==1.16.0
soupsieve==2.4
SQLAlchemy==1.3.10
typing_extensions==4.4.0
urllib3==1.25.11
Werkzeug==2.0.0
WTForms==3.0.1
zipp==3.11.0
